requirements.txt
streamlit
pandas
openai
python-dotenv
pypdf
python-docx
Pillow
gspread
oauth2client